Extract CellIssueComponent from SheetIssuesComponent
